Improve Solo search discoverability
