+++ frontend/app/sitemap.ts
@@ -1,0 +1,11 @@
+import type { MetadataRoute } from "next";
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  return [
+    {
+      url: "https://soloagent.team/",
+      changeFrequency: "weekly",
+      priority: 1,
+    },
+  ];
+}
